Rename SeqViewerComponent to SeqViz
- refactor so annotations without names are accepted
