+++ src/SeqViz/Linear/Linear.jsx
