+++ src/SeqViz/Linear/Linear.scss
