+++ src/SeqViz/Linear/SeqBlock/Selection/Selection.jsx
