+++ src/SeqViz/Linear/SeqBlock/SeqBlock.jsx
