+++ src/SeqViz/Linear/SeqBlock/Translations/Translations.jsx
