+++ src/SeqViz/elementsToRows.js
