+++ src/SeqViz/handlers/selectionHandler.jsx
